# Improve Parser Performance