--- a/bin/create-php-parser.js
+++ b/bin/create-php-parser.js
@@ -10,6 +10,7 @@ const peg = fs.readFileSync( 'blocks/api/post.pegjs', 'utf8' );
 const parser = pegjs.generate(
 	peg,
 	{
+		cache: true,
 		plugins: [ phpegjs ],
 		phpegjs: {
 			parserNamespace: null,