--- a/lib/parser.php
+++ b/lib/parser.php
@@ -1032,46 +1032,46 @@ class Gutenberg_PEG_Parser {
           $s3 = $s4;
         }
         if ($s3 === $this->peg_FAILED) {
-          $s3 = $this->peg_currPos;
-          $s4 = $this->peg_currPos;
-          $s5 = $this->peg_currPos;
-          $this->peg_silentFails++;
-          $s6 = $this->peg_parseBlock_End();
-          $this->peg_silentFails--;
-          if ($s6 === $this->peg_FAILED) {
-            $s5 = null;
-          } else {
-            $this->peg_currPos = $s5;
-            $s5 = $this->peg_FAILED;
-          }
-          if ($s5 !== $this->peg_FAILED) {
-            if ($this->input_length > $this->peg_currPos) {
-              $s6 = $this->input_substr($this->peg_currPos, 1);
-              $this->peg_currPos++;
+          $s3 = $this->peg_parseBlock();
+          if ($s3 === $this->peg_FAILED) {
+            $s3 = $this->peg_currPos;
+            $s4 = $this->peg_currPos;
+            $s5 = $this->peg_currPos;
+            $this->peg_silentFails++;
+            $s6 = $this->peg_parseBlock_End();
+            $this->peg_silentFails--;
+            if ($s6 === $this->peg_FAILED) {
+              $s5 = null;
             } else {
-              $s6 = $this->peg_FAILED;
-              if ($this->peg_silentFails === 0) {
-                  $this->peg_fail($this->peg_c2);
-              }
+              $this->peg_currPos = $s5;
+              $s5 = $this->peg_FAILED;
             }
-            if ($s6 !== $this->peg_FAILED) {
-              $s5 = array($s5, $s6);
-              $s4 = $s5;
+            if ($s5 !== $this->peg_FAILED) {
+              if ($this->input_length > $this->peg_currPos) {
+                $s6 = $this->input_substr($this->peg_currPos, 1);
+                $this->peg_currPos++;
+              } else {
+                $s6 = $this->peg_FAILED;
+                if ($this->peg_silentFails === 0) {
+                    $this->peg_fail($this->peg_c2);
+                }
+              }
+              if ($s6 !== $this->peg_FAILED) {
+                $s5 = array($s5, $s6);
+                $s4 = $s5;
+              } else {
+                $this->peg_currPos = $s4;
+                $s4 = $this->peg_FAILED;
+              }
             } else {
               $this->peg_currPos = $s4;
               $s4 = $this->peg_FAILED;
             }
-          } else {
-            $this->peg_currPos = $s4;
-            $s4 = $this->peg_FAILED;
-          }
-          if ($s4 !== $this->peg_FAILED) {
-            $s3 = $this->input_substr($s3, $this->peg_currPos - $s3);
-          } else {
-            $s3 = $s4;
-          }
-          if ($s3 === $this->peg_FAILED) {
-            $s3 = $this->peg_parseBlock();
+            if ($s4 !== $this->peg_FAILED) {
+              $s3 = $this->input_substr($s3, $this->peg_currPos - $s3);
+            } else {
+              $s3 = $s4;
+            }
           }
         }
         while ($s3 !== $this->peg_FAILED) {
@@ -1171,46 +1171,46 @@ class Gutenberg_PEG_Parser {
             $s3 = $s4;
           }
           if ($s3 === $this->peg_FAILED) {
-            $s3 = $this->peg_currPos;
-            $s4 = $this->peg_currPos;
-            $s5 = $this->peg_currPos;
-            $this->peg_silentFails++;
-            $s6 = $this->peg_parseBlock_End();
-            $this->peg_silentFails--;
-            if ($s6 === $this->peg_FAILED) {
-              $s5 = null;
-            } else {
-              $this->peg_currPos = $s5;
-              $s5 = $this->peg_FAILED;
-            }
-            if ($s5 !== $this->peg_FAILED) {
-              if ($this->input_length > $this->peg_currPos) {
-                $s6 = $this->input_substr($this->peg_currPos, 1);
-                $this->peg_currPos++;
+            $s3 = $this->peg_parseBlock();
+            if ($s3 === $this->peg_FAILED) {
+              $s3 = $this->peg_currPos;
+              $s4 = $this->peg_currPos;
+              $s5 = $this->peg_currPos;
+              $this->peg_silentFails++;
+              $s6 = $this->peg_parseBlock_End();
+              $this->peg_silentFails--;
+              if ($s6 === $this->peg_FAILED) {
+                $s5 = null;
               } else {
-                $s6 = $this->peg_FAILED;
-                if ($this->peg_silentFails === 0) {
-                    $this->peg_fail($this->peg_c2);
-                }
+                $this->peg_currPos = $s5;
+                $s5 = $this->peg_FAILED;
               }
-              if ($s6 !== $this->peg_FAILED) {
-                $s5 = array($s5, $s6);
-                $s4 = $s5;
+              if ($s5 !== $this->peg_FAILED) {
+                if ($this->input_length > $this->peg_currPos) {
+                  $s6 = $this->input_substr($this->peg_currPos, 1);
+                  $this->peg_currPos++;
+                } else {
+                  $s6 = $this->peg_FAILED;
+                  if ($this->peg_silentFails === 0) {
+                      $this->peg_fail($this->peg_c2);
+                  }
+                }
+                if ($s6 !== $this->peg_FAILED) {
+                  $s5 = array($s5, $s6);
+                  $s4 = $s5;
+                } else {
+                  $this->peg_currPos = $s4;
+                  $s4 = $this->peg_FAILED;
+                }
               } else {
                 $this->peg_currPos = $s4;
                 $s4 = $this->peg_FAILED;
               }
-            } else {
-              $this->peg_currPos = $s4;
-              $s4 = $this->peg_FAILED;
-            }
-            if ($s4 !== $this->peg_FAILED) {
-              $s3 = $this->input_substr($s3, $this->peg_currPos - $s3);
-            } else {
-              $s3 = $s4;
-            }
-            if ($s3 === $this->peg_FAILED) {
-              $s3 = $this->peg_parseBlock();
+              if ($s4 !== $this->peg_FAILED) {
+                $s3 = $this->input_substr($s3, $this->peg_currPos - $s3);
+              } else {
+                $s3 = $s4;
+              }
             }
           }
         }