## Description

The PHP parser has some performance issues on larger posts, I've been doing a bit of tweaking to get some more speed out of it.

The `Block_List` `pre` rule now does a simple search for strings that aren't `<!--`. Its a reasonable assumption that the vast majority of strings that do match this will be the start of a block, so we're probably going to the `bs` rule once it does match.

The `Block_Balanced` `children` rule has a similar simple search, for the same reason.

The `Block_Name_Part` rule explicitly defines a handful of block names. This feels a bit weird to be doing in the parser, but `$( [a-z][a-z0-9_-]* )` is such a slow rule to match, that I think it's worth it.

I've also enabled pegjs' internal caching on the PHP parser, as this appear to give a decent boost, too.

## How has this been tested?

Comparing the current parser with this PR, I get some pretty clear performance gains:

| File | Current | This PR | Improvement |
| --- | --- | --- | --- |
| [demo-post.html](https://raw.githubusercontent.com/dmsnell/gutenberg-document-library/master/library/demo-post.html) | 29.8ms | 7.3ms | 4x |
| [shortcode-shortcomings.html](https://raw.githubusercontent.com/dmsnell/gutenberg-document-library/master/library/shortcode-shortcomings.html) | 77.9ms | 16.5ms | 4x |
| [redesigning-chrome-desktop.html](https://raw.githubusercontent.com/dmsnell/gutenberg-document-library/master/library/redesigning-chrome-desktop.html) | 227.9ms | 52.6ms | 4x |
| [web-at-maximum-fps.html](https://raw.githubusercontent.com/dmsnell/gutenberg-document-library/master/library/web-at-maximum-fps.html) | 169.1ms | 41.3ms | 4x |
| [early-adopting-the-future.html](https://raw.githubusercontent.com/dmsnell/gutenberg-document-library/master/library/early-adopting-the-future.html) | 270.8ms | 62.9ms | 4x |
| [pygmalian-raw-html.html](https://raw.githubusercontent.com/dmsnell/gutenberg-document-library/master/library/pygmalian-raw-html.html) | 421.9ms | 254.6ms | 1x |
| [moby-dick-parsed.html](https://raw.githubusercontent.com/dmsnell/gutenberg-document-library/master/library/moby-dick-parsed.html) | 5402.3ms | 1197.7ms | 4x |