@pento 2018-07-19
Hacking this into `::parse()` (and `$prepre` handling into `peg_join_blocks()`) improves `pygmalian-raw-html.html` from a 1x increase to a 6x increase. (In raw numbers, ~420ms down to ~60ms.)

```php
	$prepre = '';
	for ( $ii = 0; $ii < $this->input_length - 3; $ii++ ) {
		if (
			'<' === $this->input[ $ii ] &&
			'!' === $this->input[ $ii + 1 ] &&
			'-' === $this->input[ $ii + 2 ] &&
			'-' === $this->input[ $ii + 3 ]
			) {
				break;
			}

			$prepre .= $this->input[ $ii ];
	}

	if ( $ii === $this->input_length - 3 ) {
		$prepre .= $this->input_length[ $ii ]
				. $this->input_length[ $ii + 1 ]
				. $this->input_length[ $ii + 2 ];

		$this->cleanup_state();
		return array( array( 'attrs' => array(), 'innerHTML' => $prepre ) );
	}

	if ( $ii > 0 ) {
		$this->input = array_slice( $this->input, $ii );
		$this->input_length -= $ii;
	}
```