@gziolo 2018-07-19
I moved `blocks/api/post.pegjs` to its own package in #7664. It generates new JS file for parser out of updated grammar every time you run `npm install`, `npm run dev` or `npm run build`. Let's make sure it gets properly regenerated with the proposed changes.