@nylen 2018-07-19
> It's a pity that the pegjs model parses an array instead of char-by-char, making this function necessary.

It sounds like this is the thing to fix here.  [phpegjs](https://github.com/nylen/phpegjs) does this to avoid issues with the ways multibyte char handling differs in JS vs PHP, see https://github.com/WordPress/gutenberg/pull/1775.

I think that when the grammar rules themselves do not contain multibyte chars (like the Gutenberg grammar), we should be able to work directly with a string instead.  How can we test and prove that?

> I've also enabled pegjs' internal caching on the PHP parser, as this appear to give a decent boost, too.

How does this change affect memory usage during parsing?