@dmsnell 2018-07-19
Thanks for working on this @pento.

I think that this will end up failing the `Block_List` won't it if we have a normal HTML comment somewhere before the first block?

I've got mixed feelings too about optimizing the grammar itself. I know we have need but I'd rather see us build a new implementation of the PHP parser by hand which conforms to the grammar. We _do_ have ways of automating the tests and performance of alternate parsers.

Since the grammar isn't all that complicated I think we can make a good one by hand that is faster than the generated one. I'd love it if we can keep the _intention_ and _specification_ as clear as they can be.