@pento 2018-07-20
> I think that when the grammar rules themselves do not contain multibyte chars (like the Gutenberg grammar), we should be able to work directly with a string instead. How can we test and prove that?

Yah, that should be possible. UTF-8 encoding requires that only ASCII characters can have the `0xxxxxxx` bit structure, all other characters (including single-byte UTF-8 characters) must be made of bytes that have variations of the `1xxxxxxx` bit structure. So, it's not possible for an ASCII character to match any part of a UTF-8 character.

> I think that this will end up failing the Block_List won't it if we have a normal HTML comment somewhere before the first block?

Maybe, I didn't test it that much. 🙂

> I've got mixed feelings too about optimizing the grammar itself. I know we have need but I'd rather see us build a new implementation of the PHP parser by hand which conforms to the grammar.

This is another viable option if the [PCRE parser](https://github.com/Hywan/gutenberg-parser-rs/blob/php_fallback/bindings/php/pure/ParserPCRE.php) doesn't work out. 

I tend to agree that, rather than optimising the grammar, a better focus would be on improving `phpegjs` performance for cases that are valuable to us, though that gets a bit tricky with the current grammar. For example, `$("<!--" .)*` could pretty clearly be optimised to use `strpos()` to find the next occurrence of `"<!--"`, grab everything that happened before it, and move the `currPos` forward. Optimising `$(!Block .)*`, on the other hand, which could benefit from exactly the same optimisation, would be significantly more involved.

Hand-coding a PHP parser is an option that would probably give us better memory usage than PCRE, but I don't know how well it would do on time. It would also be less than fun to maintain, if we were to make significant changes to the grammar.

Anyway, thanks for the thoughts, y'all. I'm going to close this for now, pending the outcome of the significantly more advanced experiments in PCRE parsing.